+++ client/src/app/modules/pixlisecore/components/atoms/selection-changer/selection-changer.component.ts
@@ -156,7 +156,11 @@ export class SelectionChangerComponent implements OnInit, OnDestroy {
   }
 
   onClear(): void {
-    this._selectionService.clearSelection();
+    const allScanIds = [];
+    for (const scan of Object.values(this._analysisLayoutService.activeScreenConfiguration$.value.scanConfigurations)) {
+      allScanIds.push(scan.id);
+    }
+    this._selectionService.clearSelection(allScanIds);
   }
 
   private getSelectionScanIds(): string[] {